requirements.txt
black
boto3
botocore
click
jmespath
joblib
mypy-extensions
numpy
packaging
pandas
pathspec
platformdirs
python-dateutil
pytz
PyYAML
pylint
black
autopep8
s3transfer
scikit-learn
scipy
six
threadpoolctl
tzdata
urllib3

